ModeSlice: Refactor to keep list of valid modes
This change allows us to set which modes are valid for a given game mode.
This is useful as the Player has three valid entry modes (Value, Corner,
Center) but the Builder only has one (Value).

Previously, we passed a builder flag to the GameArea.  If it was set,
space would not trigger a nextMode dispatch.
Also we hid the change modes buttons in Builder's Controls.

There were some issues with the old approach
- Non-descriptive prop names. The behavior to not change values was tied to
  the a prop name 'builder'.  If I were fixing instead of changing old
  approach I would rename the prop something like 'allowModeChange'
- Inflexibility.  The old approach was all-or-nothing. With the new
  approach we would be able to-- for instance-- allow the builder to set
  both Values and Corner marks, but not Center marks
- Lack of guarantees.  The old approach provided no guarantees that a
  nextMode or setMode action would not be dispatched.  If such a dispatch
  were somehow overlooked the application would enter an invalid state

The new approach is slightly more cumbersome (and requires testing) but
should help with the issues outlined above